training and exercise logic developed

+++ src/main/java/com/example/elgrande/repository/UserRepository.java
@@ -2,7 +2,7 @@ package com.example.elgrande.repository;
 
 import com.example.elgrande.model.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-    // Add custom query methods if needed
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
